[ci, sival] Add azure CI testplan lint job on PR
Add an Azure CI job to run on a pull request that will ensure the testplans are formatted correctly by using a corresponding linter script, currently set to run on both the `hw/top_earlgrey/data` and `hw/top_earlgrey/data/ip` directories. This is also added to `quick_lint` under `ci/jobs/quick-lint.sh` so that it can be run locally as a quick linter.

Signed-off-by: AlexJones0 <alex.jones@lowrisc.org>