[i2c, dif] Test cycle bounds for clock stretching in i2c dif
The i2c block requires a minimum value (for current setup, at least 4 cycles) for the SCL high and low time in order for clock stretching to function. This functionality was previously added to the DIF, and this commit updates the relevant unit tests to also check for these conditional cases. The hard-coded parameter of 4 cycles has been extracted to the header so it can be referenced in a static assert before the test, and the tests themselves modify the input clock period in order to cause the DIF to calculate a case that will execute the minimum boundary branches.

Signed-off-by: Alex Jones <alex.jones@lowrisc.org>