[bazel,ci] Enable target-location.sh to process multiple file outputs.
This commit first switches target-location.sh from using bazel `outquery` to bazel `cquery --output=files` so that in the case of an
`opentitan_binary` software target with multiple potential outputs we can find any of them using the script. I use `cquery` and
`grep` rather than `outquery-all` or `outquery-${BOARD}.bin` to avoid existing OpenTitan/Issue#23990.

We then ensure that the execution root is added to all output files rather than just the first, so that all queried files are
translated to absolute paths rather than relative paths.

A `rgrep target-location.sh` has been used to confirm that the places this  script was being used in already (in the `azure-pipelines.yml`
to find the Test ROM for the cw305, and in `ci/scripts/get-bitstream-strategy.sh` to get the location of the bitstream manifest file)
are unaffected by this change.

Signed-off-by: Alex Jones <alex.jones@lowrisc.org>