[docs,bazel] Update documentation to correct FPGA bootstrapping process
This commit first adds the `fpga_cw340` execution environment to the hello_world demo example `opentitan_binary`, so that
the demo can be built and run on the CW340 like is advertised in the docs.

Documentation files are then also updated to provide commands to properly build and locate this binary in accordance
with the current usage of commands. This commit also fixes a couple of semantic errors in the documentation, and provides
a couple sentences of extra information on `opentitan_test`, `opentitan_binary` and troubleshooting that a new user may
find useful.

Signed-off-by: Alex Jones <alex.jones@lowrisc.org>