[bazel] Ensure that the output files of outquery are unique
This commit resolves OpenTitan/Issue#23990, ensuring that the output of bazel outquery/outquery-one is unique.
This is because some scripts rely on the fact that outquery will only return a single file corresponding to some
unique software target in certain cases, but this may not be the case when an `opentitan_test` is built using
`./bazelisk.sh build ...`. The target is built in multiple configurations and thus bazel `cquery` returns the
same file multiple times, which is unexpected behaviour.

A better solution in the future would be to replace `--output=starlark ...` with `--output=files` later on
as this enforces uniqueness internal to Bazel, without any extra processing on top, but this requires replacing
existing post-processing via starlark expressions for different forms of `./bazelisk.sh outquery`. This ideally
should be replaced though as it was added before the `--output=files` functionality was introduced to Bazel,
which is much more suitable for the use case here.

Signed-off-by: Alex Jones <alex.jones@lowrisc.org>