[sw,lib] Add unit test for debug printing FourCC types
Adds unit tests for debug formatting of the `kFourCC` type added in
OpenTitan/PR#22198. This formatting specifier style ("%C") prints a uint32_t as
four ASCII characters encoded in little endian order. Existing print unit tests
did not check this formatting specifier, and so this commit introduces three
additional small unit tests to check that this implementation is working as was
intended.

Signed-off-by: Alex Jones <alex.jones@lowrisc.org>