[sw,dv] Add basic support for `status_t` (`%r`) logging in DV
Adds a simple substitution rule to the current C->SV logging backdoor (which
was introduced to bypass long UART times for logging during DV simulation).
This rule replaces the "%r" and "%!r" formatting specifiers (prints a `status_t`
normally and in a JSON representation respectively) with a hex dump of the
32 bit packed `status_t` value.

This is implemented to prevent errors when using `status_t` in DV tests. A full
formatting implementation is not made due to the large amounts of code duplication
required - a far better solution that still supports the C->SV backdoor would be
to use DPI to run the C logging code in SV, but that involves some more intricate
and complicated processes that shold be considered in more detail in a separate
issue. This commit provides a temporary fix until that implementation is in place.

Signed-off-by: Alex Jones <alex.jones@lowrisc.org>