[entropy_src, verilator] Reduce contiguous samples on Verilator
Reduces the number of contiguous entropy samples observed when running
on Verilator to just 8 (down from 1024). This is because observing the
full 1024 would require increasing the timeout greatly and would mean
that it would take around 13-15 hours to run the test. Without
increasing the timeout, the test currently fails due to not meeting the
required rate.

Signed-off-by: Alex Jones <alex.jones@lowrisc.org>