[entropy_src, verilator] Restrict consumers to random nums on Verilator
This commit modifies the `entroy_src_fw_override_test` such that when
running on Verilator, the `firmware_override_extract_insert` test
function will only execute random number generation as an entropy
consumer, and will simply return before executing any other entropy
consumers.

Entropy is generated much more slowly on Verilator and thus this
would otherwise take an impractical amount of time, even if timeout was
greatly increased (and options for timeout existed). Instead, since
we don't care about the rate of entropy consumption on Verilator, we
can restrict the test to just generating random numbers to verify that
the entropy fw override is functioning as we would expect.

This commit also extracts a few left-over constant values into named
enum values for readability, and increases the timeout to generate
random numbers on Verilator (for the above reasons) so that the test
can succeed as expected.

Signed-off-by: Alex Jones <alex.jones@lowrisc.org>