[dttool] Handle special `INOUT_A0` port_type for EG IOA2/3
In `top_earlgrey.hjson`, the IOA2 and IOA3 pads have their port type
overriden from the default `inout wire` to the value `INOUT_A0.
According to the relevant comment, this is only used for simulating ADC
connections right now, where these ports are switched to a real type
when the SystemVerilog ANALOGSIM macro is defined (i.e. this would work
if we parsed a pre-processed SV file, but not when generating from this
HJSON).

Since this appears to be the one unique case of a macro being used for a
port-type in a hacky way, we just hard-code this into the relevant lists
in dtgen/helper.py.

Signed-off-by: Alex Jones <alex.jones@lowrisc.org>