[git] Better .gitattributes for github-linguist
Try and improve the default language detection by github-linguist to
give better and more accurate statistics for the OpenTitan repository,
as well as syntax highlighting support for a larger number of files.

This is just an improvement and not intended to be perfect; template
files mix Mako's syntax with the source and do not have a grammar that
can parse them native to Github Linguist, so we opt to just use the
generated language's grammar where possible. Some file types
(exclusions/waivers) are not recognised as languages by linguist.

Signed-off-by: Alex Jones <alex.jones@lowrisc.org>