Initial BitbangWrapperUart stub implementation
Provides the interface to dispatch a custom Uart that will wrap existing
calls with equivalent bitbangs. This has access to the parameters of the
underyling UARTs (flow control, parity, baud rate etc.) and its own
software-provided `Vec` buffers for RX and TX operation. The pins that
will be used for GPIO bitbanging are computed from the configuration pin
mapping and the UART name, though none of the actual bit-banging is
implemented yet.

Signed-off-by: Alex Jones <alex.jones@lowrisc.org>