[opentitanlib] Apply default UART configuration parameters to transports
Previously, unlike SPI and I2C configurations, UART configuration
parameters (derived from the json configuration files) were not used by
opentitanlib when creating the UART objects. This commit mirrors the
existing I2C and SPI logic to propagate these parameters to the created
UART devices.

For some transports these parameters appear to be exposing the defaults
and do not seem to be meant to be used to set the values (e.g. parity /
baud rate / stop bits may not be configurable), so we ignore any errors
in setting the parameters on startup. A better long-term solution might
be to have the created UART trait expose whether these parameters are
modifiable or not (or to error only if the set value does not match the
actual implementation).

Signed-off-by: Alex Jones <alex.jones@lowrisc.org>