[opentitanlib] Add stubbed `wrap_bitbangs` TransportWrapper option
This options remains stubbed out for now but will be populated in future
commits.

The motivation for this option is to be able to wrap any arbitrary
transport that supports GPIO bitbanging and monitoring (i.e. it provides
pin manipulation & monitoring capabilities) so that, if possible, other
I/O operations such as UART and SPI are all performed directly via
software using this bit-banging interface.

The end goal is to support hardware peripherals (UART and SPI) for
backends that do not support them, but have good bitbanging
functionalities. By wrapping an arbitrary transport, we make use of
the underyling trait implementations and can run existing FPGA tests via
bitbanging to verify the generated bitbanging functionality.

Signed-off-by: Alex Jones <alex.jones@lowrisc.org>