[opentitanlib] Add UART bitbang wrapper read functionality
Implements the RX monitoring for the UART bitbanging wrapper, meaning
that when existing UART consoles are created and used in tests where
bitbang wrapping is enabled, UART reads now use an underlying GPIO
monitoring mechanism with a software decoder rather than using an actual
hardware UART.

To use the GPIO monitoring mechanism (which produces waveforms as a
sequence of rising/falling edge events) with the UART decoder (which
takes a uniformly sampled waveform) this commit introduces logic to
sample the wave at regular intervals and decode on the fly based on what
has been received so far.

From experimental testing with the hyperdebug backend, this seems
accurate up to at least 115200 bps, with slight instabilities
(occassional bitflips at 230400 bps).

Signed-off-by: Alex Jones <alex.jones@lowrisc.org>