[sival,kmac] Add kmac_mode_change_test
Adds a simple test that checks that the KMAC has the ability to
transition from being configured using EDN-sourced entropy to using
SW-sourced entropy, so long as there is a timeout on the entropy
arrival. This mechanism allows us to derisk use of the entropy complex.

There is no way for software to read the current latched entropy mode
from the KMAC Entropy FSM. However, currently, once we have switched to
SW-seeded entropy, there is no way to switch back (SW-seeded entropy
is terminal, bar resetting the IP). While this is the case, we can
exploit this functionality to test that we have transitioned to
SW-seeded entropy, by ensuring that the previous timeout condition no
longer occurs.

Signed-off-by: Alex Jones <alex.jones@lowrisc.org>