[opentitanlib] Add SPI bitbang wrapper
Implement the SPI bitbang wrapper interface, meaning that when existing
SPI peripherals are created and used in tests where bitbang wrapping is
enabled, SPI transactions (writes/reads/both) now use an underlying GPIO
bitbanging mechanism with a software encoder, rather than using an
actual hardware SPI Host.

Since we are driving the clock and SPI is synchronous, we do not have to
be as concerned with clock jitter/skew and asynchronous transmission
like we are with UART bitbanging. Using Hyperdebug, this should be able
to clock at 50kHz fine, and potentially a decent amount faster depending
on Hyperdebug's bitbanging accuracy.

Due to issues with Hyperdebug's firmware when attempting to Bitbang a
large enough number of pins (see relevant comment in the code),
dual/quad mode is not fully implemented for now and as such we use the
default dispatcher for EEPROM transactions, meaning that the bitbang
wrapper SPI might not support e.g. 1-1-4 SPI mode even though the
underlying SPI could.

Signed-off-by: Alex Jones <alex.jones@lowrisc.org>