[opentitantool] Chunk large Hyperdebug commands to avoid RX overflow
Introduce a mechanism for chunking Hyperdebug commands that are > 58
characters to avoid overflowing the Hyperdebug EC firmwares USB console
stream <-> USART host console RX buffer, which is defined as 64 bytes (a
USB packet size). Hyperdebug is defined as allowing commands of up to
256 characters, so we need to adequately chunk our commands to transmit
valid Hyperdebug commands longer than 64 bytes.

This commit tries to change the existing behaviour as little as possible
whilst introducing the additional logic to allow for this chunking, and
separating out the response parsing logic into isolated components. To
ease this implementation, the one change that is made is that for the
first command sent, the Ctrl-C transmission used to clear the existing
output is now separate from the transmission of the first command. This
should have negligible performance impact, as it should only happen
once. Otherwise, all commands < 58 bytes will transmit as before,
whilst longer commands will be chunked into multiple transmissions.

Signed-off-by: Alex Jones <alex.jones@lowrisc.org>