[opentitanlib] Add dual/quad pin config to SPI bitbang wrapper
The logic to implement dual/quad SPI read or writes is not yet present,
but this commit introduces the relevant logic for configuring the pins
for dual/quad reads and writes, including data types corresponding to
the pin configurations, which will depend on both the data mode (single,
dual, quad) and the direction of transfer (e.g. you can have dual/quad
reads and writes), such as in e.g. 1-1-2 (Dual SPI) 1-2-2 (DIO) or
4-4-4 (QPI) modes etc.

Signed-off-by: Alex Jones <alex.jones@lowrisc.org>