[flash,sival] Add flash endurance testpoint
For PROD silicon, we want a test that we can manually run to ensure that
the on-chip flash can endure a certain number of Program/Erase cycles
before it wears and starts to show faults. Add the testpoint for this
test.

Running this test will wear down the flash and so should only be done
manually, and carefully. We test the last working page so that the test
can be run multiple times on the same chip without issues, wearing down
pages that are less likely to be used by other tests or firmware.

Signed-off-by: Alex Jones <alex.jones@lowrisc.org>