[sival, flash_ctrl] Add flash endurance test
Designed to be run on silicon, but can be tested with a limited number
of cycles on other targets (e.g. FPGA).

Find the last writable page(s) in bank 1's data partition and repeatedly
program/erase until a fault is seen. Check that the number of cycles
until we see a fault (flash wear) exceeds a preset target value. Try
this both with and without high-endurance enabled.

Signed-off-by: Alex Jones <alex.jones@lowrisc.org>