[sival, flash_ctrl] Check single ECC errors in flash endurance test
Improve the flash endurance test so that, when ECC is enabled and a test
wants to, we can check for single ECC errors that are automatically
corrected by the flash controller and choose whether to count them as
errors or not.

Adds a test that uses ECC enabled (without scrambling or high
endurance), which allows single ECC errors. We might expect the ECC
correction to make the flash slightly more enduring.

Signed-off-by: Alex Jones <alex.jones@lowrisc.org>