[sival, flash_ctrl] Report detailed flash info in flash endurance test
Refactor the flash endurance test to now detect errors at a flash-word
level of granularity in both erases and programs, so that we can keep
testing when some flash words have failed but others have not (whereas
before we would stop testing the entire page when an any single flash
word faults).

Use this information to continue testing until >= 50% of the flash words
have become faulty, and compute statistics which are then reported based
on these 50% of failing words in the page. This includes reporting the
minimum cycles to failure (which we compare with the target), the
*average* cycles to failure (for the worst 50% of words), and the
*variance* in the cycles to failure.

Signed-off-by: Alex Jones <alex.jones@lowrisc.org>