[sw] Add `bitbanged_uart` test
Adds a software test that can be run in CI to check that Hyperdebug's
bitbang_wrapper functionality is working correctly for the transport
UART. This functionality will be useful for manufacturing/provisioning
workflows, and so it will be useful to have a test checking that this
keeps working in CI.

This is a very simple test that just sends and receives some data
between the Host and device at 57600 Bd, to work around Hyperdebug not
being able to handle 115200 Bd by default. It also makes minimal use of
a SPI console for synchronisation, to avoid the UART console having the
same issue by picking up the device's default baud rate.

This will allow us to test the UART bitbanging remains working (to at
least some limited degree) in CI.

Signed-off-by: Alex Jones <alex.jones@lowrisc.org>