[sw] Add `bitbanged_spi` test
Adds a software test that can be run in CI to check that Hyperdebug's
bitbang_wrapper functionality is working correctly for the SPI. This
functionality will be useful for manufacturing/provisioning workflows,
and so it will be useful to have a test checking that this keeps working
in CI.

This test just takes the existing SPI device ottf_console_test (which
only uses SPI) and wraps it so that it uses the bitbanged SPI instead.
Since there are no clock issues as the host is driving the clock, there
are no additional considerations that need to be made for this test
(unlike the UART test).

Signed-off-by: Alex Jones <alex.jones@lowrisc.org>