[flash,sival] Implement embedded flash endurance test
This commit implements the flash endurance test for both FPGA and
silicon targets, with the device acting as a command processor that
receives flash commands from the host and executes them accordingly.
The host parses the commands to be run from a JSON command file to
permit customisation of the endurance test based on the testing
environment, the vendor-specific flash macro, and the target
specifications.

See the subsequent commit for more documentation.

Signed-off-by: Alex Jones <alex.jones@lowrisc.org>