[bazel,qemu] Add QEMU envs to attestation `print_certs` test
Adds the QEMU ROM_EXT execution environment to the end-to-end ROM_EXT
attestation test that tests we can read out valid attestation data and
print the generated DICE certificates.

This test target requires bootstrapping and as such needs an appropriate
bootstrapping delay (e.g. 1 second) which is higher than the default for
QEMU environments, as a result of QEMU's slow emulation of the ROM
compared to FPGA/silicon. This is due to the SEC_MMIO hot loop in the
ROM falling in an unaligned (to a 4K page) EPMP region which causes
QEMU's TCG's TLB caching to be unused.

Signed-off-by: Alex Jones <alex.jones@lowrisc.org>