[qemu,manuf] Change `cp_provision_functest` to use GPIO 3
As a workaround for the fact that QEMU' GPIO is currently not routed via
Pinmux, change the TX Ready Indiciator GPIO (used by the SPI console) to
be GPIO 3 instead of GPIO 0, to match the GPIO that is used by the
regular `cp_provision` binary, which uses other GPIO for ATE integration.
This allows the `cp_provision_functest` to pass in the QEMU environment.

Signed-off-by: Alex Jones <alex.jones@lowrisc.org>