[manuf] Refactor CP/FT orchestrator logic to support more exec envs
Adds initial support for the QEMU execution target into the orchestrator
by refactoring the CP/FT flows to be more generic about the defined
execution target, with some refactoring to reduce common repetitions
across the FPGA and silicon targets (and simulation/emulation, in the
future).

This is currently untested for QEMU - it simply aims to refactor the
script enough that support for QEMU (and other simulation / emulation)
targets can be more easily introduced in the future.

The `--fpga` argument on the orchestrator is changed to `--exec-target`,
and now additionally accepts `qemu` and `silicon`. `silicon` remains the
default value, and the behaviour should otherwise be unchanged.

Signed-off-by: Alex Jones <alex.jones@lowrisc.org>