[bazel,qemu] Refactor QEMU rules to allow standalone execution
Refactor all of the logic for constructing the arguments that must be
passed to QEMU to run an OT Earlgrey 1.0.0 machine into a separate
`qemu.sh` utility script that can be used to spawn a standalone instance
of QEMU, without going through Bazel. This allows for more widespread
use of QEMU outside of the current `opentitan_test` environment, which it
was previously constrained to.

This commit also makes a couple of modifications to the existing flow:
 - More arguments are now parameterized and can be changed as part of
   this standalone flow (e.g. the monitor/socket paths).
 - The cleanup trap handler logic in `qemu_test.sh` is fixed to take the
   daemonized process ID returned by QEMU, rather than trying to kill
   the original QEMU process (which should exit after daemonization,
   hence it was not doing anything and processes could persist through a
   `bazel run`). The kill output is silenced to clean up some noise.

Signed-off-by: Alex Jones <alex.jones@lowrisc.org>
Co-authored-by: James Wainwright <james.wainwright@lowrisc.org>