[manuf] Fix orchestrator subprocess stdout flakiness
We need to wait for the tee outputs to finish, otherwise there may be a
final line buffered in stdout/stderr that is not yet written. In the
unlucky case (can rarely happen a few percent of the time), we try and
look for the CHIP_PROBE_DATA etc. in the stdout logfile but do not find
it because it was not written yet. This is more likely under higher
processing load. This fix removes any such flakiness from the
orchestrator.

Signed-off-by: Alex Jones <alex.jones@lowrisc.org>