[qemu,opentitanlib] Switch QEMU monitor CharDev to a socket
Improve the idempotence of connecting to the QEMU monitor CharDev with
OpenTitanLib/OpenTitanTool by using a socket instead of a PTY for the
CharDev backend.

This ensures that a new buffer will be allocated on each new
opentitanlib connection to QEMU and that no stale data will be retained
in a kernel PTY buffer. Helps resolve potential edge cases where we try
to reconnect but there is still some partial data left over in the PTY
buffer from a previous connection (e.g. after a Ctrl-C).

Signed-off-by: Alex Jones <alex.jones@lowrisc.org>