[qemu,opentitanlib] Add CharDev path override option
Add an option that can be used to override the paths of arbitrary QEMU
CharDevs (PTYs / sockets) that we connect to. The current code assumes
that the paths retrieved from the monitor are correct. But this is not
necessarily the case if QEMU was instantiated with sockets created at
relative paths and the working directory has changed.

As such, you can now override these paths by passing in arguments. For
example, you might use the options:
```
opentitantool \
    --qemu-device-path taprbb=/custom/rv_dm/jtag/path \
    --qemu-device-path uart1=/custom/uart0/path \
    --qemu-device-path gpio=/custom/gpio/path \
    /* ... other options ... */
```
to override the `taprbb`, `uart1` and `gpio` CharDevs respectively.

Signed-off-by: Alex Jones <alex.jones@lowrisc.org>