[qemu,opentitanlib] Switch QEMU GPIO CharDev to a socket
Improve the idempotence of connecting to the QEMU GPIO CharDev with
OpenTitanLib/OpenTitanTool by using a socket instead of a PTY for the
CharDev backend.

When previously using a PTY, it was very common to find problems on a
reconnection:
 - Stale partial line data in the kernel PTY buffer.
 - QEMU doesn't detect another `OPEN` event and so does not know to
   update the host with its current GPIO state.
 - PTY File descriptor not properly closed.

While these could be worked around somewhat, for accurate emulation and
synchronization we really need to switch to use a socket, so that
separate buffers are assigned to each outgoing/incoming GPIO connection
and so that QEMU can be properly notified about the host connection.

Signed-off-by: Alex Jones <alex.jones@lowrisc.org>