[bazel,qemu] Add fake signed QEMU binaries
To build the orchestrator zip, signed ROM_EXT and ft_personalize QEMU
binaries need to be present. Since there hasn't been a signing ceremony
with the addition of QEMU, these do not exist.

To allow us to run the orchestrator for non-SiVal SKUs for testing,
check in some fake signed binaries (empty files) to allow the
orchestrator zip to build for QEMU. These can later be replaced by
actual signed binaries for QEMU in a signing ceremony.

Signed-off-by: Alex Jones <alex.jones@lowrisc.org>