[orchestrator] Add QEMU support to the orchestrator
This commit directly implements QEMU support in the orchestrator,
defining the correct host flags that are needed for each of the CP/FT
stages in order to execute the tests in QEMU.

As this change increases the amount of repeated code, there is also some
cleanup/refactoring in this commit in an effort to reduce duplication
between the CP/FT stages and between different targets. Modularisation
is introduced for various types of host targets, but these should be
unchanged for the existing silicon/FPGA environments (modulo some docs),
with the only change on the QEMU exec target path.

Signed-off-by: Alex Jones <alex.jones@lowrisc.org>