[bazel,qemu] Add QEMU support for E2E orchestrator tests
This commit adds Bazel versions all of the existing E2E orchestrator
tests. We would expect all of these tests to work in QEMU with the
exception of the `e2e_sival_qemu_test`, as real SiVal binaries are not
currently checked in.

These tests already had some shared/duplicated logic which increased
when QEMU support was added. To address this, perform some small
refactoring of shared data into a common dictionary per execution target
to reduce this repetition.

Signed-off-by: Alex Jones <alex.jones@lowrisc.org>